feat: 상품 조회 쿼리 튜닝

+++ src/main/java/com/palgona/palgona/product/dto/response/ProductPageResponse.java
@@ -7,7 +7,7 @@ public record ProductPageResponse(
         Long id,
         String name,
         int currentBid,
-        int bookmarkCount,
+        Long bookmarkCount,
         LocalDateTime deadline,
         LocalDateTime created_at,
         String imageUrl
